requirements.txt
pillow
numpy
torch